requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.35
pg8000==1.31.2
google-cloud-secret-manager==2.20.2
cloud-sql-python-connector==1.18.4
